Refactor the bulbs logic into a separate file to keep text.py cleaner
